Authenticatie af

+++ app/Http/Controllers/QuizController.php
@@ -5,6 +5,9 @@ use App\Models\Quiz;
 use App\Models\Like;
 use App\Models\Tag;
 
+use Auth;
+use Gate;
+
 use Illuminate\Http\Request;
 use App\Http\Requests;
 
@@ -20,6 +23,9 @@ class QuizController extends Controller
     //alle quizzes opvragen op admin pagina
     public function getAdminIndex()
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $quizzes = Quiz::orderBy('title', 'asc')->get();
         return view('admin.index', ['quizzes' => $quizzes]);
     }
@@ -45,6 +51,9 @@ class QuizController extends Controller
     //1 quiz opvragen op admin pagina om te maken
     public function getAdminCreate()
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
@@ -52,6 +61,9 @@ class QuizController extends Controller
     //1 quiz opvragen op admin pagina om te editen  
     public function getAdminUpdate($id)
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $quiz = Quiz::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['quiz' => $quiz, 'quizId' => $id, 'tags' => $tags]);
@@ -60,14 +72,22 @@ class QuizController extends Controller
     //Admin create post opdracht afhandelen
     public function postAdminCreate(Request $request)
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [ 
             'title' => 'required|min:5'
         ]);
+        //kijken of de user ingelogd is en anders terugsturen
+        $user = Auth::user();
+        if (!$user) {
+            return redirect()->back();
+        }
         $quiz = new Quiz([
             'title' => $request->input('title'), 
             'content' => $request->input('content')
         ]);
-        $quiz->save();
+        $user->quizzes()->save($quiz);
         $quiz->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
 
         return redirect()->route('admin.index')->with('info', 'Quiz created: ' . $request->input('title'));
@@ -76,10 +96,16 @@ class QuizController extends Controller
     //Admin update post opdracht afhandelen
     public function postAdminUpdate(Request $request)
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $this->validate($request, [ 
             'title' => 'required|min:5'
         ]);
         $quiz = Quiz::find($request->input('id'));
+        if (Gate::denies('manipulate-quiz', $quiz)) {
+            return redirect()->back()->with('error', 'This quiz is not yours!');
+        }
         $quiz->title = $request->input('title');
         $quiz->content = $request->input('content');
         $quiz->save();
@@ -90,7 +116,13 @@ class QuizController extends Controller
 
     public function getAdminDelete($id)
     {
+        if(!Auth::check()) {
+            return redirect()->back();
+        }
         $quiz = Quiz::find($id);
+        if (Gate::denies('manipulate-quiz', $quiz)) {
+            return redirect()->back()->with('info', 'This quiz is not yours!');
+        }
         $quiz->likes()->delete();
         $quiz->tags()->detach();
         $quiz->delete();
